+++ src/test/scala/fastml4j/regression/LinearRegressionSuite.scala
@@ -14,19 +14,19 @@ import fastml4j.util.DataGenerators._
 class LinearRegressionSuite extends FunSuite with BeforeAndAfter {
 
   test("Simple test") {
-    val coef1 = 1.0
-    val coef2 = 0.5
+    val coef1 = 1.0f
+    val coef2 = 0.5f
 
-    val x = (0 to   500).map { x => Array(x.toDouble,1.0) }.toArray
+    val x = (0 to   500).map { x => Array(x.toFloat,1.0) }.toArray
     val y = x.map {case Array(x1, x2) => Array(x1 * coef1 + x2 * coef2 + math.random/10) }
     val dataSet = new DataSet(x.toNDArray, y.toNDArray)
 
-    val lr = new LinearRegression(lambdaL2 = 0.0, alpha = 0.000001, eps = 1e-4, maxIterations = 2000)
+    val lr = new LinearRegression(lambdaL2 = 0.0f, alpha = 0.000001f, eps = 1e-4f, maxIterations = 2000)
     lr.fit(dataSet)
 
     val weights = lr.weights
-    assert(weights.getDouble(0,0) === coef1 +- 1)
-    assert(weights.getDouble(0,1) === coef2 +- 1)
+    assert(weights.getFloat(0,0) === coef1 +- 1)
+    assert(weights.getFloat(0,1) === coef2 +- 1)
   }
 /*
   test("with data generators") {
@@ -40,8 +40,8 @@ class LinearRegressionSuite extends FunSuite with BeforeAndAfter {
     lr.fit(dataset)
     val weights = lr.weights
     println(weights)
-    assert(weights.getDouble(0,0) === coef1 +- 0.1)
-    assert(weights.getDouble(0,1) === coef2 +- 0.1)
+    assert(weights.getFloat(0,0) === coef1 +- 0.1)
+    assert(weights.getFloat(0,1) === coef2 +- 0.1)
 
 
   }*/
